+++ docs/changes.rst
@@ -14,6 +14,7 @@ Not yet released.
 * Superuser access to components with :ref:`component-restricted`.
 * Adjusted default :setting:`LOGIN_REQUIRED_URLS_EXCEPTIONS` to not block :ref:`manage-appearance`.
 * Avoid crash on pushing changes to diverged repository.
+* Avoid crash when installing :ref:`addon-weblate.generate.pseudolocale`.
 
 **Compatibility**